Mehs loading changes

+++ Hazard-Renderer/src/Backend/Metal/MTLUtils.h
@@ -120,7 +120,7 @@ namespace HazardRenderer::Metal
     }
     uint32_t GetMipLevelCount(uint32_t width, uint32_t height)
     {
-        return (uint32_t)std::floor(std::log2(glm::min(width, height))) + 1;
+        return std::floor(std::log2(glm::min(width, height))) + 1;
     }
 }
 #endif
